Raise WindowDragger PositionChangeStart/End on the UI thread
The WindowDragger receives a callback from the LL MouseHook. We cannot
process outbound RPC calls from within that thread. However, the
callbcak for the PositionChangeEnd can call into the VDS service to
check whether the window has been pinned, which is a COM call.

This causes the process to crash. To avoid this, raise the events
asynchronously.