Use KeyCode instead of Windows.Input.Key
KeyCode corresponds to the Win32 Virtual Key  codes. Windows.Input.Key is .NET specific.

Plan is to unify the hotkey handling logic, so use KeyCode as the more natural representation everywhere.

Note: Due to inconsistent naming between the two enums, and having to update that, which affects the parsing logic, the Activation Hotkey will be reset to the default after upgrade.