Add automatic split trigger
Automatically creates a split once the parent contains more than the
configured number of nodes.

The default is set to 2. This changes the default behaviour for all
users, but can be reverted by setting the threshold to a high value.

With the default behaviour, opening a sequence of windows results in a
spiral layout with 2 nodes at each level, which has been highly
requested.

In my testing, I've found the new default to be more intuitive and to
almost always achieve what I would have done manually. For people on
wide-screen displays, tweaking the threshold to 3 might allow for better
screen space utilisation.