Fix invalid node swapping causing stack overflow
When using the mouse + swap (Shift key), it is possible to (incorrectly)
cause FancyWM to swap a node with it's parent in the internal tree
structure. This operation leaves the SplitPanel with a child pointer
which points to itself.

This can result in a StackOverflow, when the Window traverses the Parent
pointers to reach the root desktop node.